* Major tweak to the serialiser: shifted size to 32bits to handle bigger packets.  This will break compatibility.  * Modified tlvitems to support future expansions.  * added random tlv test functios.  * added dummy Wide Directory functions to rsdir (for the future;)  * Fixed Mutex Bug in fimonitor.  * added netiface_test for hiberation / net change testing.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@1620 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ libretroshare/src/serialiser/rstlvkeys.h
@@ -48,7 +48,7 @@ class RsTlvSecurityKey: public RsTlvItem
 	public:
 	 RsTlvSecurityKey();
 virtual ~RsTlvSecurityKey() { return; }
-virtual uint16_t TlvSize();
+virtual uint32_t TlvSize();
 virtual void	 TlvClear();
 virtual bool     SetTlv(void *data, uint32_t size, uint32_t *offset); /* serialise   */
 virtual bool     GetTlv(void *data, uint32_t size, uint32_t *offset); /* deserialise */
@@ -69,7 +69,7 @@ class RsTlvSecurityKeySet: public RsTlvItem
 	public:
 	 RsTlvSecurityKeySet() { return; }
 virtual ~RsTlvSecurityKeySet() { return; }
-virtual uint16_t TlvSize();
+virtual uint32_t TlvSize();
 virtual void	 TlvClear();
 virtual bool     SetTlv(void *data, uint32_t size, uint32_t *offset); /* serialise   */
 virtual bool     GetTlv(void *data, uint32_t size, uint32_t *offset); /* deserialise */
@@ -85,7 +85,7 @@ class RsTlvKeySignature: public RsTlvItem
 	public:
 	 RsTlvKeySignature();
 virtual ~RsTlvKeySignature() { return; }
-virtual uint16_t TlvSize();
+virtual uint32_t TlvSize();
 virtual void	 TlvClear();
 virtual bool     SetTlv(void *data, uint32_t size, uint32_t *offset); /* serialise   */
 virtual bool     GetTlv(void *data, uint32_t size, uint32_t *offset); /* deserialise */
